pyproject.toml
[tool.poetry]
name = "django-unmigrate"
version = "0.1.4"
description = "Smart reversion of Django migrations based on Git diff"
authors = ["Lorenzo Peña <lorinkoz@gmail.com>"]
license = "MIT"
readme = "README.rst"
repository = "https://github.com/lorinkoz/django-unmigrate"
keywords = ["django", "git", "migrations"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django",
    "Framework :: Django :: 2.0",
    "Framework :: Django :: 2.1",
    "Framework :: Django :: 2.2",
    "Framework :: Django :: 3.0",
]

[tool.poetry.dependencies]
python = "^3.6"
django = ">=2.0,<4.0"
gitpython = "^3.1.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
coverage = {extras = ["toml"], version = "^5.1"}
flake8 = "^3.8.4"
flake8-bugbear = "^20.11.1"
flake8-comprehensions = "^3.3.1"
flake8-no-types = "^1.1.1"
flake8-tidy-imports = "^4.2.1"
isort = "^5.7.0"

[tool.black]
line-length = 120

[tool.isort]
atomic = true
combine_as_imports = true
line_length = 120
profile = "black"

[tool.coverage.run]
source = ["django_unmigrate"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
